reservation-service/requirements.txt
fastapi
uvicorn
motor
pydantic
grpcio
grpcio-tools
pika
pytest
httpx
pymongo
protobuf
anyio
asgiref
pytest-asyncio
python-dotenv
